+++ middleware/recovery/recover.go
@@ -1,23 +1,23 @@
-package recover
+package recovery
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/botchris/go-pubsub"
 )
 
-// RecoveryHandlerFunc is a function that recovers from the panic `p` by returning an `error`.
-type RecoveryHandlerFunc func(ctx context.Context, p interface{}) error
+// HandlerFunc is a function that recovers from the panic `p` by returning
+// an `error` message
+type HandlerFunc func(ctx context.Context, p interface{}) error
 
 type middleware struct {
 	pubsub.Broker
-	handler RecoveryHandlerFunc
+	handler HandlerFunc
 }
 
 // NewRecoveryMiddleware returns a new middleware that recovers from panics when
 // publishing to topics or delivering to subscribers.
-func NewRecoveryMiddleware(parent pubsub.Broker, handler RecoveryHandlerFunc) pubsub.Broker {
+func NewRecoveryMiddleware(parent pubsub.Broker, handler HandlerFunc) pubsub.Broker {
 	return &middleware{
 		Broker:  parent,
 		handler: handler,
@@ -26,8 +26,8 @@ func NewRecoveryMiddleware(parent pubsub.Broker, handler RecoveryHandlerFunc) pu
 
 func (mw middleware) Publish(ctx context.Context, topic pubsub.Topic, m interface{}) (err error) {
 	defer func(ctx context.Context) {
-		if r := recover(); r != nil {
-			err = recoverFrom(ctx, r, "pubsub: publisher panic\n", mw.handler)
+		if p := recover(); p != nil {
+			err = mw.handler(ctx, p)
 		}
 	}(ctx)
 
@@ -47,12 +47,4 @@ func (mw middleware) Subscribe(ctx context.Context, topic pubsub.Topic, sub pubs
 
 func (mw middleware) Unsubscribe(ctx context.Context, topic pubsub.Topic, subscriber pubsub.Subscriber) error {
 	return mw.Broker.Unsubscribe(ctx, topic, subscriber)
-}
-
-func recoverFrom(ctx context.Context, p interface{}, wrap string, r RecoveryHandlerFunc) error {
-	if r == nil {
-		return fmt.Errorf("%s: %s", p, wrap)
-	}
-
-	return r(ctx, p)
 }
